+++ .gitignore
@@ -1,3 +1,5 @@
+# needs fixing
+Makefile
 
 # Random list of yt links that now lives inside this repo folder for some reason...
 yt*.txt
