+++ src/scripts/hack/hack.js
