+++ src/scripts/hack/manager.js
