+++ src/scripts/hack/start-automate.js
