+++ src/scripts/hack/stop-automate.js
