+++ src/scripts/hack/weaken.js
