+++ src/scripts/helpers/global-params.js
