+++ src/scripts/shop/server-rack.js
