--- src/scripts/test-dir.ts
@@ -1,5 +1,0 @@
-import type { NS } from "../../NetscriptDefinitions.d.ts"
-
-export async function main(ns: NS): Promise<void> {
-  ns.tprint(ns.getHostname());
-}
